requirements.txt
asgiref==3.7.2
asttokens==2.4.1
autopep8==2.0.4
build==1.0.3
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.3
dj-database-url==2.1.0
Django==4.1.13
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-modeltranslation==0.18.11
django-split-settings==1.2.0
djangorestframework==3.14.0
executing==2.0.1
fire==0.5.0
gunicorn==21.2.0
icecream==2.1.3
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==23.2
pillow==10.2.0
pip-tools==7.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
pyproject_hooks==1.0.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
rav==0.0.9
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
termcolor==2.4.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
